+++ packages/memo/src/index.ts
@@ -191,7 +191,7 @@ export function createWritableMemo<T>(
   options?: MemoOptions<T | undefined>,
 ): [signal: Accessor<T>, setter: Setter<T>] {
   let combined: Accessor<T> = () => value as T;
-  const [signal, setSignal] = createSignal(value as T),
+  const [signal, setSignal] = createSignal(value as T, EQUALS_FALSE_OPTIONS),
     memo = createMemo(
       callbackWith(fn, () => combined()),
       value,