+++ packages/pagination/src/index.ts
@@ -1,10 +1,4 @@
-import {
-  access,
-  tryOnCleanup,
-  type Directive,
-  noop,
-  type MaybeAccessor,
-} from "@solid-primitives/utils";
+import { access, tryOnCleanup, noop, type MaybeAccessor } from "@solid-primitives/utils";
 import {
   Accessor,
   batch,
@@ -270,7 +264,7 @@ export type _E = JSX.Element;
  */
 export function createInfiniteScroll<T>(fetcher: (page: number) => Promise<T[]>): [
   pages: Accessor<T[]>,
-  loader: Directive,
+  loader: (el: Element) => void,
   options: {
     page: Accessor<number>;
     setPage: Setter<number>;