+++ packages/resource/src/index.ts
@@ -5,19 +5,22 @@ import {
   type ResourceFetcher,
   type ResourceFetcherInfo,
   type Signal,
+  onCleanup,
 } from "solid-js";
 import { createStore, reconcile, unwrap } from "solid-js/store";
 
 export type AbortableOptions = {
-  noAutoAbort?: true;
+  noAutoAbort?: boolean;
   timeout?: number;
 };
 
 /**
  * **Creates and handles an AbortSignal**
  * ```ts
- * const [signal, abort] = makeAbortable({ timeout: 10000 });
- * const fetcher = (url) => fetch(url, { signal: signal() });
+ * const [signal, abort, filterAbortError] =
+ *   makeAbortable({ timeout: 10000 });
+ * const fetcher = (url) => fetch(url, { signal: signal() })
+ *   .catch(filterAbortError); // filters abort errors
  * ```
  * Returns an accessor for the signal and the abort callback.
  *
@@ -26,25 +29,59 @@ export type AbortableOptions = {
  * - `noAutoAbort`: can be set to true to make a new source not automatically abort a previous request
  */
 export function makeAbortable(
-  options?: AbortableOptions,
-): [signal: () => AbortSignal, abort: () => void] {
+  options: AbortableOptions = {},
+): [
+  signal: () => AbortSignal,
+  abort: (reason?: string) => void,
+  filterAbortError: (err: any) => void,
+] {
   let controller: AbortController | undefined;
   let timeout: NodeJS.Timeout | number | undefined;
-  const abort = () => {
+  const abort = (reason?: string) => {
     timeout && clearTimeout(timeout);
-    controller?.abort();
+    controller?.abort(reason);
   };
   const signal = () => {
-    if (!options?.noAutoAbort && controller?.signal.aborted === false) {
-      abort();
+    if (!options.noAutoAbort && controller?.signal.aborted === false) {
+      abort("retry");
     }
     controller = new AbortController();
-    if (options?.timeout) {
-      timeout = setTimeout(abort, options.timeout);
+    if (options.timeout) {
+      timeout = setTimeout(() => abort("timeout"), options.timeout);
     }
     return controller.signal;
   };
-  return [signal, abort];
+  return [
+    signal,
+    abort,
+    err => {
+      if (err.name === "AbortError") {
+        return undefined;
+      }
+      throw err;
+    },
+  ];
+}
+
+/**
+ * **Creates and handles an AbortSignal with automated cleanup**
+ * ```ts
+ * const [signal, abort, filterAbortError] =
+ *   createAbortable();
+ * const fetcher = (url) => fetch(url, { signal: signal() })
+ *   .catch(filterAbortError); // filters abort errors
+ * ```
+ * Returns an accessor for the signal and the abort callback.
+ *
+ * Options are optional and include:
+ * - `timeout`: time in Milliseconds after which the fetcher aborts automatically
+ * - `noAutoAbort`: can be set to true to make a new source not automatically abort a previous request
+ */
+
+export function createAbortable(options?: AbortableOptions) {
+  const [signal, abort, filterAbortError] = makeAbortable(options);
+  onCleanup(abort);
+  return [signal, abort, filterAbortError];
 }
 
 const mapEntries = (entries: [key: string, value: any][]) =>